Adding task 12

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -22,5 +22,5 @@ void print_number(int n)
 	{
 		print_number(c);
 	}
-	_putchar(n % 10 + '0');
+	_putchar((unsigned int) n % 10 + '0');
 }
